renaming jpegs to jpgs, test if that works with DB
